+++ changelog.md
@@ -5,6 +5,7 @@ v 1.1.3
 - Added support for PHP in views. Depends on file extension.
 - Support now multiple database engines for the moment PDO and RedBeans (This is still subject to change).
 - Widget approach support.
+- Fixed problem in sub-folder resolution in the controller folder.
 
 v 1.1.2